removed everything from starting_params.py that has nothing to do with starting parameters.
