+++ lib/spectrum_analysis/mapping.py
@@ -4,7 +4,9 @@ import matplotlib
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 
 from spectrum_analysis.spectrum import *
+from spectrum_analysis.dictionaries import *
 from spectrum_analysis import data
+
 
 """
 This module contains the mapping class to work multiple x, y structured
@@ -443,7 +445,7 @@ class mapping(spectrum):
             peaktype = re.sub('_p.*_', '', peak)
 
             # create model with parameters as before
-            model = ChoosePeakType(peaktype, number)
+            model = self.ChoosePeakType(peaktype, number)
             model = StartingParameters(model, peaks)
             model.make_params()
 
@@ -757,27 +759,6 @@ class mapping(spectrum):
     def PlotAllColormaps(self, maptype, y, mapdims, step, **kwargs):
         """
         """
-        cmaps = [('Perceptually Uniform Sequential', [
-            'viridis', 'plasma', 'inferno', 'magma']),
-         ('Sequential', [
-            'Greys', 'Purples', 'Blues', 'Greens', 'Oranges', 'Reds',
-            'YlOrBr', 'YlOrRd', 'OrRd', 'PuRd', 'RdPu', 'BuPu',
-            'GnBu', 'PuBu', 'YlGnBu', 'PuBuGn', 'BuGn', 'YlGn']),
-         ('Sequential (2)', [
-            'binary', 'gist_yarg', 'gist_gray', 'gray', 'bone', 'pink',
-            'spring', 'summer', 'autumn', 'winter', 'cool', 'Wistia',
-            'hot', 'afmhot', 'gist_heat', 'copper']),
-         ('Diverging', [
-            'PiYG', 'PRGn', 'BrBG', 'PuOr', 'RdGy', 'RdBu',
-            'RdYlBu', 'RdYlGn', 'Spectral', 'coolwarm', 'bwr', 'seismic']),
-         ('Qualitative', [
-            'Pastel1', 'Pastel2', 'Paired', 'Accent',
-            'Dark2', 'Set1', 'Set2', 'Set3',
-            'tab10', 'tab20', 'tab20b', 'tab20c']),
-         ('Miscellaneous', [
-            'flag', 'prism', 'ocean', 'gist_earth', 'terrain', 'gist_stern',
-            'gnuplot', 'gnuplot2', 'CMRmap', 'cubehelix', 'brg', 'hsv',
-            'gist_rainbow', 'rainbow', 'jet', 'nipy_spectral', 'gist_ncar'])]
         for category in cmaps:
             print(category[0])
             for colormap in category[1]:
